Add files via upload
This is the start of the redesign - it is very loose right now so I am currently working on adding documentation and creating a README. The biggest task list item as of right now is finalizing and testing the view class - making sure that everything displays as it should. Then, I will go back and connect the internal slideshow and menu management to the view.